+++ packages/synapse-interface/components/_Transaction/_Transaction.tsx
@@ -6,7 +6,6 @@ import { getTxBlockExplorerLink } from './helpers/getTxBlockExplorerLink'
 import { getExplorerAddressLink } from './helpers/getExplorerAddressLink'
 import { useBridgeTxStatus } from './helpers/useBridgeTxStatus'
 import { isNull } from 'lodash'
-import { DownArrow } from '../icons/DownArrow'
 import {
   updateTransactionKappa,
   completeTransaction,
@@ -17,7 +16,9 @@ import { Chain, Token } from '@/utils/types'
 import TransactionArrow from '../icons/TransactionArrow'
 import { TimeRemaining } from './components/TimeRemaining'
 import { TransactionStatus } from './components/TransactionStatus'
-import { useEstimatedTimeStatus } from './helpers/useEstimatedTimeStatus'
+import { getEstimatedTimeStatus } from './helpers/getEstimatedTimeStatus'
+import { DropdownMenu } from './components/DropdownMenu'
+import { MenuItem } from './components/DropdownMenu'
 
 interface _TransactionProps {
   connectedAddress: string
@@ -67,7 +68,7 @@ export const _Transaction = ({
     elapsedTime: estimatedElapsedTime,
     startCheckingTimeReached,
     remainingTimeInMinutes: EstimatedRemainingTimeInMinutes,
-  } = useEstimatedTimeStatus(currentTime, timestamp, estimatedTime)
+  } = getEstimatedTimeStatus(currentTime, timestamp, estimatedTime)
 
   const elapsedTime: number = currentTime - timestamp // in seconds
   const remainingTime: number = estimatedTime - elapsedTime
@@ -226,84 +227,5 @@ export const _Transaction = ({
         </div>
       </div>
     </div>
-  )
-}
-
-export const DropdownMenu = ({ children }) => {
-  const [open, setOpen] = useState<boolean>(false)
-
-  const handleClick = () => {
-    setOpen(!open)
-  }
-
-  return (
-    <div className="relative">
-      <div
-        onClick={handleClick}
-        className={`
-          rounded w-5 h-[21px] flex place-items-center justify-center
-          bg-surface
-          border border-solid border-[--synapse-border]
-          hover:border-[--synapse-focus]
-          cursor-pointer
-        `}
-      >
-        <DownArrow />
-      </div>
-
-      {open && (
-        <ul
-          className={`
-            absolute z-50 mt-1 p-0 bg-surface border border-solid border-tint rounded shadow popover -right-1 list-none text-left text-sm
-          `}
-        >
-          {children}
-        </ul>
-      )}
-    </div>
-  )
-}
-
-export const MenuItem = ({
-  text,
-  link,
-  onClick,
-}: {
-  text: string
-  link: string
-  onClick?: () => any
-}) => {
-  return (
-    <li
-      className={`
-      rounded cursor-pointer
-      border border-solid border-transparent
-      hover:border-[--synapse-focus]
-      active:opacity-40
-    `}
-    >
-      {onClick ? (
-        <div
-          onClick={onClick}
-          className={`
-            block pl-2 pr-3 py-2 whitespace-nowrap text-[--synapse-text-primary] no-underline after:content-['_↗'] after:text-xs after:text-[--synapse-secondary]
-          `}
-        >
-          {text}
-        </div>
-      ) : (
-        <a
-          href={link ?? ''}
-          onClick={onClick}
-          target="_blank"
-          rel="noreferrer"
-          className={`
-            block pl-2 pr-3 py-2 whitespace-nowrap text-[--synapse-text-primary] no-underline after:content-['_↗'] after:text-xs after:text-[--synapse-secondary]
-          `}
-        >
-          {text}
-        </a>
-      )}
-    </li>
   )
 }
